+++ packages/apis/example/lib/services/handlers/orders_handlers/order_handlers/retrieves_get_orders_authorized_handler.dart
@@ -1,10 +1,10 @@
 import 'package:apis/apis.dart';
-import 'package:apis/network/remote/orders/order/abstract/order.dart';
+import 'package:apis/network/remote/orders/order/abstract/order_service.dart';
 import 'package:example/services/api_request_handler.dart';
 import 'package:example/services/api_service_registry.dart';
 import 'package:get_it/get_it.dart';
 
-class GetOrdersWithPropertiesHandler implements ApiRequestHandler {
+class GetOrdersAuthorizedHandler implements ApiRequestHandler {
   @override
   Future<Map<String, dynamic>> handleRequest(
       String method, Map<String, String> params) async {
@@ -23,14 +23,9 @@ class GetOrdersWithPropertiesHandler implements ApiRequestHandler {
     final String? updatedAtMax = params['updated_at_max'];
     final String? processedAtMin = params['processed_at_min'];
     final String? processedAtMax = params['processed_at_max'];
-    final String? attributionAppId = params['attribution_app_id'];
-    final String? status = params['status'];
-    final String? financialStatus = params['financial_status'];
-    final String? fulfillmentStatus = params['fulfillment_status'];
-    final String? fields = params['fields'];
 
     try {
-      final response = await GetIt.I<OrderService>().getOrdersWithProperties(
+      final response = await GetIt.I<OrderService>().getOrdersAuthorized(
         apiVersion: apiVersion,
         createdAtMin: createdAtMin,
         createdAtMax: createdAtMax,
@@ -38,23 +33,18 @@ class GetOrdersWithPropertiesHandler implements ApiRequestHandler {
         updatedAtMax: updatedAtMax,
         processedAtMin: processedAtMin,
         processedAtMax: processedAtMax,
-        attributionAppId: attributionAppId,
-        status: status,
-        financialStatus: financialStatus,
-        fulfillmentStatus: fulfillmentStatus,
-        fields: fields,
       );
 
       return {
         "status": "success",
-        "message": "Orders fetched successfully",
+        "message": "Authorized orders fetched successfully",
         "orders": response.orders?.map((order) => order.toJson()).toList(),
         "timestamp": DateTime.now().toIso8601String(),
       };
     } catch (e) {
       return {
         "status": "error",
-        "message": "Failed to fetch orders: ${e.toString()}",
+        "message": "Failed to fetch authorized orders: ${e.toString()}",
         "timestamp": DateTime.now().toIso8601String(),
       };
     }
@@ -70,11 +60,6 @@ class GetOrdersWithPropertiesHandler implements ApiRequestHandler {
 
   Map<String, List<ApiField>> get optionalFields => {
         'GET': [
-          const ApiField(
-            name: 'api_version',
-            label: 'API Version',
-            hint: 'Enter the Shopify API version (e.g., 2024-04)',
-          ),
           const ApiField(
             name: 'created_at_min',
             label: 'Created At Min',
@@ -106,29 +91,34 @@ class GetOrdersWithPropertiesHandler implements ApiRequestHandler {
             hint: 'Filter orders processed before this timestamp (ISO 8601)',
           ),
           const ApiField(
-            name: 'attribution_app_id',
-            label: 'Attribution App ID',
-            hint: 'Filter orders by the app that created them',
+            name: 'created_at_min',
+            label: 'Created At Min',
+            hint: 'Filter orders created after this timestamp (ISO 8601)',
           ),
           const ApiField(
-            name: 'status',
-            label: 'Status',
-            hint: 'Filter by order status',
+            name: 'created_at_max',
+            label: 'Created At Max',
+            hint: 'Filter orders created before this timestamp (ISO 8601)',
           ),
           const ApiField(
-            name: 'financial_status',
-            label: 'Financial Status',
-            hint: 'Filter by financial status',
+            name: 'updated_at_min',
+            label: 'Updated At Min',
+            hint: 'Filter orders updated after this timestamp (ISO 8601)',
           ),
           const ApiField(
-            name: 'fulfillment_status',
-            label: 'Fulfillment Status',
-            hint: 'Filter by fulfillment status',
+            name: 'updated_at_max',
+            label: 'Updated At Max',
+            hint: 'Filter orders updated before this timestamp (ISO 8601)',
           ),
           const ApiField(
-            name: 'fields',
-            label: 'Fields',
-            hint: 'Comma-separated list of fields to include in the response',
+            name: 'processed_at_min',
+            label: 'Processed At Min',
+            hint: 'Filter orders processed after this timestamp (ISO 8601)',
+          ),
+          const ApiField(
+            name: 'processed_at_max',
+            label: 'Processed At Max',
+            hint: 'Filter orders processed before this timestamp (ISO 8601)',
           ),
         ],
       };
