+++ packages/apis/example/lib/services/handlers/orders_handlers/order_handlers/update_email_address_handler.dart
@@ -1,12 +1,11 @@
 import 'package:apis/apis.dart';
-import 'package:apis/network/remote/orders/order/abstract/order.dart';
-import 'package:apis/network/remote/orders/order/freezed_model/request/update_phone_number_request.dart';
+import 'package:apis/network/remote/orders/order/abstract/order_service.dart';
+import 'package:apis/network/remote/orders/order/freezed_model/request/update_email_address_request.dart';
 import 'package:example/services/api_request_handler.dart';
 import 'package:example/services/api_service_registry.dart';
-
 import 'package:get_it/get_it.dart';
 
-class UpdatePhoneNumberHandler implements ApiRequestHandler {
+class UpdateEmailAddressHandler implements ApiRequestHandler {
   @override
   List<String> get supportedMethods => ['PUT'];
 
@@ -20,9 +19,9 @@ class UpdatePhoneNumberHandler implements ApiRequestHandler {
             isRequired: true,
           ),
           const ApiField(
-            name: 'phone',
-            label: 'Phone Number',
-            hint: 'New phone number for the order',
+            name: 'email',
+            label: 'Email Address',
+            hint: 'New email address for the order',
             isRequired: true,
           ),
         ],
@@ -31,7 +30,7 @@ class UpdatePhoneNumberHandler implements ApiRequestHandler {
   @override
   Future<Map<String, dynamic>> handleRequest(
       String method, Map<String, String> params) async {
-    // Only handle PUT requests for updating phone number
+    // Only handle PUT requests for updating email address
     if (method != 'PUT') {
       return {
         "status": "error",
@@ -43,7 +42,7 @@ class UpdatePhoneNumberHandler implements ApiRequestHandler {
     // Extract required parameters
     final String apiVersion = params['api_version'] ?? ApiNetwork.apiVersion;
     final orderId = params['order_id'];
-    final phone = params['phone'];
+    final email = params['email'];
 
     // Validate required parameters
     if (orderId == null || orderId.isEmpty) {
@@ -54,25 +53,25 @@ class UpdatePhoneNumberHandler implements ApiRequestHandler {
       };
     }
 
-    if (phone == null || phone.isEmpty) {
+    if (email == null || email.isEmpty) {
       return {
         "status": "error",
-        "message": "Phone number is required",
+        "message": "Email address is required",
         "timestamp": DateTime.now().toIso8601String(),
       };
     }
 
     try {
       // Create the request model
-      final model = UpdatePhoneNumberRequest(
+      final model = UpdateEmailAddressRequest(
         order: Order(
           id: int.tryParse(orderId),
-          phone: phone,
+          email: email,
         ),
       );
 
       // Call the order service
-      final response = await GetIt.I<OrderService>().updatePhoneNumber(
+      final response = await GetIt.I<OrderService>().updateEmailAddress(
         apiVersion: apiVersion,
         orderId: orderId,
         model: model,
@@ -81,7 +80,7 @@ class UpdatePhoneNumberHandler implements ApiRequestHandler {
       // Return success response
       return {
         "status": "success",
-        "message": "Order phone number updated successfully",
+        "message": "Order email address updated successfully",
         "order": response.order?.toJson(),
         "timestamp": DateTime.now().toIso8601String(),
       };
@@ -89,7 +88,7 @@ class UpdatePhoneNumberHandler implements ApiRequestHandler {
       // Handle errors
       return {
         "status": "error",
-        "message": "Failed to update order phone number: ${e.toString()}",
+        "message": "Failed to update order email address: ${e.toString()}",
         "timestamp": DateTime.now().toIso8601String(),
       };
     }
