+++ packages/apis/example/lib/services/handlers/orders_handlers/order_handlers/update_note_attributes_handler.dart
@@ -1,15 +1,15 @@
 import 'package:apis/apis.dart';
-import 'package:apis/network/remote/orders/order/abstract/order.dart';
-import 'package:apis/network/remote/orders/order/freezed_model/request/update_remove_customer_order_request.dart';
+import 'package:apis/network/remote/orders/order/abstract/order_service.dart';
+import 'package:apis/network/remote/orders/order/freezed_model/request/update_note_attributes_request.dart';
 import 'package:example/services/api_request_handler.dart';
 import 'package:example/services/api_service_registry.dart';
 import 'package:get_it/get_it.dart';
 
-class UpdateRemoveCustomerHandler implements ApiRequestHandler {
+class UpdateNoteAttributesHandler implements ApiRequestHandler {
   @override
   Future<Map<String, dynamic>> handleRequest(
       String method, Map<String, String> params) async {
-    // Only handle PUT requests for removing customer from order
+    // Only handle PUT requests for updating note attributes
     if (method != 'PUT') {
       return {
         "status": "error",
@@ -21,6 +21,8 @@ class UpdateRemoveCustomerHandler implements ApiRequestHandler {
     // Extract required parameters
     final String apiVersion = params['api_version'] ?? ApiNetwork.apiVersion;
     final orderId = params['order_id'];
+    final name = params['note_attribute_name'];
+    final value = params['note_attribute_value'];
 
     // Validate required parameters
     if (orderId == null || orderId.isEmpty) {
@@ -31,17 +33,30 @@ class UpdateRemoveCustomerHandler implements ApiRequestHandler {
       };
     }
 
+    if (name == null || name.isEmpty || value == null || value.isEmpty) {
+      return {
+        "status": "error",
+        "message": "Note attribute name and value are required",
+        "timestamp": DateTime.now().toIso8601String(),
+      };
+    }
+
     try {
       // Create the request model
-      final model = UpdateRemoveCustomerOrderRequest(
+      final model = UpdateNoteAttributesRequest(
         order: Order(
           id: int.tryParse(orderId),
-          customer: null, // Setting customer to null removes it from the order
+          noteAttributes: [
+            NoteAttribute(
+              name: name,
+              value: value,
+            ),
+          ],
         ),
       );
 
       // Call the order service
-      final response = await GetIt.I<OrderService>().updateRemoveCustomer(
+      final response = await GetIt.I<OrderService>().updateNoteAttributes(
         apiVersion: apiVersion,
         orderId: orderId,
         model: model,
@@ -50,15 +65,15 @@ class UpdateRemoveCustomerHandler implements ApiRequestHandler {
       // Return success response
       return {
         "status": "success",
-        "message": "Customer removed from order successfully",
-        "order": response.order?.toJson(),
+        "message": "Note attributes updated successfully",
+        "order": response.toJson(),
         "timestamp": DateTime.now().toIso8601String(),
       };
     } catch (e) {
       // Handle errors
       return {
         "status": "error",
-        "message": "Failed to remove customer from order: ${e.toString()}",
+        "message": "Failed to update note attributes: ${e.toString()}",
         "timestamp": DateTime.now().toIso8601String(),
       };
     }
@@ -73,7 +88,19 @@ class UpdateRemoveCustomerHandler implements ApiRequestHandler {
           const ApiField(
             name: 'order_id',
             label: 'Order ID',
-            hint: 'The ID of the order to remove customer from',
+            hint: 'The ID of the order to update',
+            isRequired: true,
+          ),
+          const ApiField(
+            name: 'note_attribute_name',
+            label: 'Note Attribute Name',
+            hint: 'The name of the note attribute',
+            isRequired: true,
+          ),
+          const ApiField(
+            name: 'note_attribute_value',
+            label: 'Note Attribute Value',
+            hint: 'The value of the note attribute',
             isRequired: true,
           ),
         ],
