+++ packages/apis/example/lib/services/handlers/orders_handlers/order_handlers/update_remove_customer_handler.dart
@@ -1,36 +1,15 @@
 import 'package:apis/apis.dart';
-import 'package:apis/network/remote/orders/order/abstract/order.dart';
-import 'package:apis/network/remote/orders/order/freezed_model/request/update_email_address_request.dart';
+import 'package:apis/network/remote/orders/order/abstract/order_service.dart';
+import 'package:apis/network/remote/orders/order/freezed_model/request/update_remove_customer_order_request.dart';
 import 'package:example/services/api_request_handler.dart';
 import 'package:example/services/api_service_registry.dart';
 import 'package:get_it/get_it.dart';
 
-class UpdateEmailAddressHandler implements ApiRequestHandler {
-  @override
-  List<String> get supportedMethods => ['PUT'];
-
-  @override
-  Map<String, List<ApiField>> get requiredFields => {
-        'PUT': [
-          const ApiField(
-            name: 'order_id',
-            label: 'Order ID',
-            hint: 'The ID of the order to update',
-            isRequired: true,
-          ),
-          const ApiField(
-            name: 'email',
-            label: 'Email Address',
-            hint: 'New email address for the order',
-            isRequired: true,
-          ),
-        ],
-      };
-
+class UpdateRemoveCustomerHandler implements ApiRequestHandler {
   @override
   Future<Map<String, dynamic>> handleRequest(
       String method, Map<String, String> params) async {
-    // Only handle PUT requests for updating email address
+    // Only handle PUT requests for removing customer from order
     if (method != 'PUT') {
       return {
         "status": "error",
@@ -42,7 +21,6 @@ class UpdateEmailAddressHandler implements ApiRequestHandler {
     // Extract required parameters
     final String apiVersion = params['api_version'] ?? ApiNetwork.apiVersion;
     final orderId = params['order_id'];
-    final email = params['email'];
 
     // Validate required parameters
     if (orderId == null || orderId.isEmpty) {
@@ -53,25 +31,17 @@ class UpdateEmailAddressHandler implements ApiRequestHandler {
       };
     }
 
-    if (email == null || email.isEmpty) {
-      return {
-        "status": "error",
-        "message": "Email address is required",
-        "timestamp": DateTime.now().toIso8601String(),
-      };
-    }
-
     try {
       // Create the request model
-      final model = UpdateEmailAddressRequest(
+      final model = UpdateRemoveCustomerOrderRequest(
         order: Order(
           id: int.tryParse(orderId),
-          email: email,
+          customer: null, // Setting customer to null removes it from the order
         ),
       );
 
       // Call the order service
-      final response = await GetIt.I<OrderService>().updateEmailAddress(
+      final response = await GetIt.I<OrderService>().updateRemoveCustomer(
         apiVersion: apiVersion,
         orderId: orderId,
         model: model,
@@ -80,7 +50,7 @@ class UpdateEmailAddressHandler implements ApiRequestHandler {
       // Return success response
       return {
         "status": "success",
-        "message": "Order email address updated successfully",
+        "message": "Customer removed from order successfully",
         "order": response.order?.toJson(),
         "timestamp": DateTime.now().toIso8601String(),
       };
@@ -88,9 +58,24 @@ class UpdateEmailAddressHandler implements ApiRequestHandler {
       // Handle errors
       return {
         "status": "error",
-        "message": "Failed to update order email address: ${e.toString()}",
+        "message": "Failed to remove customer from order: ${e.toString()}",
         "timestamp": DateTime.now().toIso8601String(),
       };
     }
   }
+
+  @override
+  List<String> get supportedMethods => ['PUT'];
+
+  @override
+  Map<String, List<ApiField>> get requiredFields => {
+        'PUT': [
+          const ApiField(
+            name: 'order_id',
+            label: 'Order ID',
+            hint: 'The ID of the order to remove customer from',
+            isRequired: true,
+          ),
+        ],
+      };
 }
