+++ packages/apis/lib/network/remote/orders/order/api/api_order_service.dart
@@ -1,6 +1,6 @@
 import 'package:apis/apis.dart';
 import 'package:apis/dio_config/api_dio_client.dart';
-import 'package:apis/network/remote/orders/order/abstract/order.dart';
+import 'package:apis/network/remote/orders/order/abstract/order_service.dart';
 import 'package:apis/network/remote/orders/order/freezed_model/request/create_cancel_order_request.dart';
 import 'package:apis/network/remote/orders/order/freezed_model/request/create_close_order_request.dart';
 import 'package:apis/network/remote/orders/order/freezed_model/request/create_order_comprehensive_request.dart';
